pyproject.toml
[tool.poetry]
name = "dugs"
version = "0.1.0"
description = ""
authors = ["DLCHAMP <36091350+dlchamp@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.8.1"
coloredlogs = "^15.0.1"
sqlalchemy = "^2.0.15"
aiosqlite = "^0.19.0"
psutil = "^5.9.5"
tabulate = "^0.9.0"
thefuzz = "^0.19.0"
python-levenshtein = "^0.21.1"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
alembic = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
